Downgrade `xstate` dependency to version 4.6.4
The current version of the visualizer requires
`xstate/lib/graph` which is not available in newer versions
of `xstate` anymore but has been moved to `@xstate/graph`.

That new version doesn't contain the `getEdges` export anymore
and appears to be generally incompatible with this version of
the visualizer.

There are probably even more incompatibilities with current
versions of `xstate`.